+++ src/albums/albums.controller.ts
@@ -9,26 +9,16 @@ import {
   UseInterceptors,
   ParseUUIDPipe,
   HttpCode,
-  Inject,
-  forwardRef,
 } from '@nestjs/common';
 import { AlbumsService } from './albums.service';
 import { CreateAlbumDto } from './dto/create-album.dto';
 import { UpdateAlbumDto } from './dto/update-album.dto';
 import { ResponseInterceptor } from '../response.interceptor';
-import { FavoritesService } from '../favorites/favorites.service';
-import { TracksService } from '../tracks/tracks.service';
 
 @Controller('album')
 @UseInterceptors(ResponseInterceptor)
 export class AlbumsController {
-  constructor(
-    @Inject(forwardRef(() => TracksService))
-    private readonly tracksService: TracksService,
-    @Inject(forwardRef(() => FavoritesService))
-    private readonly favoritesService: FavoritesService,
-    private readonly albumsService: AlbumsService,
-  ) {}
+  constructor(private readonly albumsService: AlbumsService) {}
 
   @Post()
   create(@Body() createAlbumDto: CreateAlbumDto) {
@@ -56,8 +46,6 @@ export class AlbumsController {
   @Delete(':id')
   @HttpCode(204)
   async remove(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
-    await this.favoritesService.removeAlbum(id);
-    await this.tracksService.removeAlbumId(id);
     return this.albumsService.remove(id);
   }
 }
