+++ src/artists/artists.controller.ts
@@ -9,29 +9,16 @@ import {
   UseInterceptors,
   ParseUUIDPipe,
   HttpCode,
-  Inject,
-  forwardRef,
 } from '@nestjs/common';
 import { ArtistsService } from './artists.service';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
 import { ResponseInterceptor } from '../response.interceptor';
-import { FavoritesService } from '../favorites/favorites.service';
-import { TracksService } from '../tracks/tracks.service';
-import { AlbumsService } from '../albums/albums.service';
 
 @Controller('artist')
 @UseInterceptors(ResponseInterceptor)
 export class ArtistsController {
-  constructor(
-    @Inject(forwardRef(() => FavoritesService))
-    private readonly favoritesService: FavoritesService,
-    @Inject(forwardRef(() => TracksService))
-    private readonly tracksService: TracksService,
-    @Inject(forwardRef(() => AlbumsService))
-    private readonly albumsService: AlbumsService,
-    private readonly artistsService: ArtistsService,
-  ) {}
+  constructor(private readonly artistsService: ArtistsService) {}
 
   @Post()
   create(@Body() createArtistDto: CreateArtistDto) {
@@ -59,9 +46,6 @@ export class ArtistsController {
   @Delete(':id')
   @HttpCode(204)
   async remove(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
-    await this.favoritesService.removeArtist(id);
-    this.albumsService.removeArtistId(id);
-    this.tracksService.removeArtistId(id);
     return this.artistsService.remove(id);
   }
 }
