+++ src/tracks/tracks.controller.ts
@@ -9,23 +9,16 @@ import {
   ParseUUIDPipe,
   HttpCode,
   UseInterceptors,
-  Inject,
-  forwardRef,
 } from '@nestjs/common';
 import { TracksService } from './tracks.service';
 import { CreateTrackDto } from './dto/create-track.dto';
 import { UpdateTrackDto } from './dto/update-track.dto';
 import { ResponseInterceptor } from '../response.interceptor';
-import { FavoritesService } from '../favorites/favorites.service';
 
 @Controller('track')
 @UseInterceptors(ResponseInterceptor)
 export class TracksController {
-  constructor(
-    @Inject(forwardRef(() => FavoritesService))
-    private readonly favoritesService: FavoritesService,
-    private readonly tracksService: TracksService,
-  ) {}
+  constructor(private readonly tracksService: TracksService) {}
 
   @Post()
   create(@Body() createTrackDto: CreateTrackDto) {
@@ -52,8 +45,7 @@ export class TracksController {
 
   @Delete(':id')
   @HttpCode(204)
-  async remove(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
-    await this.favoritesService.removeTrack(id);
+  remove(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
     return this.tracksService.remove(id);
   }
 }
